images functionality has been implemented and iterated on.  current code has these notes: Notes from latest build
Drag and Drop no longer works

Select images button no longer triggers cursor change and is no longer interactive

Please put a select all checkbox back in the Select header To the Left of the word Select

Please put headers above the columns in the products table like those in the images table

    Select | Product Name

Please have a select all button to the left of select in the products table header.
select checkboxes should be centered in their cells in the products table
product title should be registered left in their cells in the products table

The Product Reload action in the pulldown above the products table should be called "Reload Products"

There should not be a horizontal scrollbar on the images table window.  Make the Products Column 35% and the images column 65%.  Truncate the filename and add ellipses if the image window isn't wide enough

change the time format for Uploaded images column to 24_09_02 1:25pm

link the thumbnails in the image table to the preview image.

Please return full unabridged codeblocks for adjusted files with full verbose comments and no placeholders